+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 - [x] Auto-update preview window while typing
 - [x] Sync scroll position
-- [ ] Enter key maintains left whitespace matching previous line
+- [x] Enter key maintains left whitespace matching previous line
 - [ ] Hotkeys for h1-6 using _Ctrl+1-6_
 - [ ] Button in editor to _Copy as HTML_
 - [ ] Support for custom stylesheets